bug fixed

+++ Dto/EmployeeAttendanceSummaryDto.cs
@@ -4,7 +4,7 @@
     {
         public string EmployeeName { get; set; }
         public string EmployeeCode { get; set; }
-        public double TotalScheduledHours { get; set; }
+        public int TotalScheduledHours { get; set; }
         public double TotalWorkedHours { get; set; }
         public double TotalShortHours { get; set; }
         public double TotalExcessHours { get; set; }
